backend/requirements.txt
fastapi==0.135.1
pydantic==2.12.4
starlette==0.52.1
uvicorn==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.1